+++ Bullet.hh
@@ -9,7 +9,7 @@ class Bullet : public Entity {
 	private:
 		bool isEnemy;
 	public:
-		Bullet(Renderer *Renderer_, int x_, int y_, int vx_, int vy_, bool isEnemy);
+		Bullet(Renderer *Renderer_, int x_, int y_, float vx_, float vy_, bool isEnemy);
 		~Bullet(){delete texture;}
 		void update();
 		bool colliding(int x_, int y_){(void)x_;(void)y_;return false;}
